requirements.txt
Babel==2.3.4
Django==1.9.4
Django-Select2==4.3.2
Pillow==3.2.0
URLObject==2.4.0
Unidecode==0.04.19
YURL==0.13
aldryn-apphooks-config==0.2.7
aldryn-boilerplates==0.7.4
aldryn-categories==1.0.3
aldryn-common==1.0.1
aldryn-newsblog==1.2.1
aldryn-people==1.2.0
aldryn-reversion==1.0.8
aldryn-search==0.2.11
aldryn-translation-tools==0.2.1
backport-collections==0.1
cmsplugin-filer==1.0.1
django-appconf==1.0.2
django-appdata==0.1.5
django-bootstrap3==7.0.1
django-classy-tags==0.7.2
django-cms==3.2.5
django-compressor==2.0
django-debug-toolbar==1.4
django-dotenv==1.4.1
django-extensions==1.6.7
django-filer==1.2.0
django-filter==0.13.0
django-formtools==1.0
django-guardian==1.4.4
django-haystack==2.4.1
django-meta==1.2
django-meta-mixin==0.3.0
django-model-utils==2.5
django-mptt==0.8.4
django-parler==1.6.3
django-phonenumber-field==1.1.0
django-polymorphic==0.9.2
django-reversion==1.10.2
django-sekizai==0.9.0
django-sortedm2m==1.2.2
django-spurl==0.6.4
django-standard-form==1.1.1
django-stored-messages==1.4.0
django-taggit==0.18.3
django-taggit-autosuggest==0.2.8
django-taggit-templatetags==0.2.5
django-templatetag-sugar==1.0
django-treebeard==4.0.1
djangocms-admin-style==1.1.1
djangocms-apphook-setup==0.1.2
djangocms-blog==0.7.0
djangocms-file==1.0
djangocms-flash==0.3.0
djangocms-googlemap==0.4.0
djangocms-inherit==0.2.1
djangocms-link==1.7.2
djangocms-page-meta==0.5.11
djangocms-picture==1.0.0
djangocms-teaser==0.2.0
djangocms-text-ckeditor==2.9.3
djangocms-video==1.0.0
easy-thumbnails==2.3
html5lib==0.9999999
lxml==3.6.0
model-mommy==1.2.6
phonenumbers==7.4.0
phonenumberslite==7.4.0
psycopg2==2.6.1
pycryptodome==3.4
pylibmc==1.5.1
python-dateutil==2.5.3
python-slugify==1.2.0
pytz==2016.4
rcssmin==1.0.6
requests==2.10.0
rjsmin==1.0.12
six==1.10.0
sqlparse==0.1.19
stripe==1.33.0
wheel==0.29.0
django-admin-honeypot==1.0.0
coverage==4.3.4
git+https://github.com/ungleich/python-oca.git#egg=python-oca
djangorestframework==3.6.3
flake8==3.3.0
python-memcached==1.58
celery==4.0.2
redis==2.10.5
django-celery-results==1.0.1
kombu==4.1.0
mccabe==0.6.1
pycodestyle==2.3.1
pyflakes==1.5.0
billiard==3.5.0.3
amqp==2.2.1
vine==1.1.4